classpath move to avoid cycle
