+++ src/com/facebook/buck/android/AndroidBinary.java
@@ -23,7 +23,7 @@ import com.facebook.buck.android.FilterResourcesStep.ResourceFilter;
 import com.facebook.buck.android.ResourcesFilter.ResourceCompressionMode;
 import com.facebook.buck.io.ProjectFilesystem;
 import com.facebook.buck.java.AccumulateClassNamesStep;
-import com.facebook.buck.java.Classpaths;
+import com.facebook.buck.classpaths.Classpaths;
 import com.facebook.buck.java.HasClasspathEntries;
 import com.facebook.buck.java.JavaLibrary;
 import com.facebook.buck.java.Keystore;
