+++ src/com/facebook/buck/android/BUCK
@@ -117,6 +117,7 @@ java_immutables_library(
     ':steps',
     ':utils',
     '//src/com/facebook/buck/android/aapt:aapt',
+    '//src/com/facebook/buck/classpaths:classpaths',
     '//src/com/facebook/buck/cli:config',
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/cxx:rules',
