+++ src/com/facebook/buck/groovy/BUCK
@@ -3,6 +3,7 @@ java_library(
   srcs = glob(['*.java']),
   deps = [
     '//src/com/facebook/buck/android:packageable',
+    '//src/com/facebook/buck/classpaths:classpaths',
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/java/abi:protocol',
