+++ src/com/facebook/buck/java/BUCK
@@ -138,20 +138,17 @@ java_library(
   ],
   visibility = ['PUBLIC'],
 )
-
 java_library(
   name = 'rules',
   srcs = [
     'BuiltInJavac.java',
     'CalculateAbi.java',
-    'Classpaths.java',
     'DefaultJavaLibrary.java',
     'GwtModule.java',
     'JarFattener.java',
     'JavaBinary.java',
     'JavaBinaryDescription.java',
     'JavaLibraryDescription.java',
-    'JavaLibraryClasspathProvider.java',
     'JavaLibraryRules.java',
     'JavaSourceJar.java',
     'JavaTest.java',
@@ -178,6 +175,7 @@ java_library(
     ':steps',
     ':support',
     '//src/com/facebook/buck/android:packageable',
+    '//src/com/facebook/buck/classpaths:classpaths',
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/cxx:rules',
     '//src/com/facebook/buck/event:event',
@@ -189,25 +187,26 @@ java_library(
     '//src/com/facebook/buck/maven:util',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:rule_pattern',
+    '//src/com/facebook/buck/rules/keys:keys',
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/rules:rules',
-    '//src/com/facebook/buck/rules/keys:keys',
-    '//src/com/facebook/buck/test:test',
     '//src/com/facebook/buck/shell:steps',
-    '//src/com/facebook/buck/step:step',
     '//src/com/facebook/buck/step/fs:fs',
+    '//src/com/facebook/buck/step:step',
+    '//src/com/facebook/buck/test/result/type:type',
     '//src/com/facebook/buck/test/selectors:selectors',
+    '//src/com/facebook/buck/test:test',
     '//src/com/facebook/buck/util:constants',
     '//src/com/facebook/buck/util:exceptions',
     '//src/com/facebook/buck/util:util',
     '//src/com/facebook/buck/zip:steps',
-    '//src/com/facebook/buck/test/result/type:type',
     '//third-party/java/guava:guava',
     '//third-party/java/infer-annotations:infer-annotations',
     '//third-party/java/jsr:jsr305',
   ],
   visibility = ['PUBLIC'],
 )
+
 
 java_library(
   name = 'steps',
