+++ test/com/facebook/buck/java/BUCK
@@ -14,6 +14,7 @@ java_library(
   ],
   deps = [
     '//src/com/facebook/buck/android:packageable',
+    '//src/com/facebook/buck/classpaths:classpaths',
     '//src/com/facebook/buck/cli:config',
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/cxx:rules',
